Convert README from Markdown to reStructuredText
reStructuredText is much more common and well-supported in the Python
community. For example, all the cool kids host their documentation on
Read The Docs which supports Sphinx docs written with reStructuredText.
Also, the Python release documentation [1] says "If you use the
reStructuredText syntax for this field, PyPI will parse it and display
an HTML output for the package home page."

[1] http://docs.python.org/release/3.1.5/distutils/uploading.html